Upgrated

+++ school.Web/Pages/Floors/List.cshtml.cs
@@ -8,28 +8,27 @@ namespace SchoolManagement.Web.Pages.Floors;
 
 public class FloorListModel : PageModel
 {
-    AppDbContext _ctx;
-    private readonly IRepository<Floor>? _floorRepository;
+    private readonly IRepository<Floor> _floorRepository;
     private readonly IRepository<School> _schoolRepository;
+    public School? School { get; set; }
+    public IEnumerable<Floor>? Floors { get; private set; }
 
-    public FloorListModel(IRepository<Floor>? floorRepository, IRepository<School> schoolRepository, AppDbContext ctx)
+    public FloorListModel(IRepository<Floor> floorRepository, IRepository<School> schoolRepository)
     {
         _floorRepository = floorRepository;
         _schoolRepository = schoolRepository;
-        _ctx = ctx;
-        Floors = _ctx.Floors.Where(f=>f.SchoolId == SchoolId);
     }
-    public School School { get; set; }
-    public IEnumerable<School> Schools { get; set; }
-    public static IEnumerable<Floor> Floors { get; private set; }
 
-    public int SchoolId { get; set; }
-
-
-    public void OnGet()
+    public IActionResult OnGet()
     {
-        SchoolId = int.Parse(HttpContext.Request.Cookies["SchoolId"]);
-        Floors = _floorRepository.GetAll(f => f.SchoolId == SchoolId);
+        var sId = HttpContext.Request.Cookies["SchoolId"];
+        if (!int.TryParse(sId, out int schoolId))
+        {
+            return NotFound("School not found");
+        }
+
+        Floors = _floorRepository!.GetAll(f => f.SchoolId == schoolId);
+        return Page();
     }
 
     public void OnPost()
@@ -37,45 +36,66 @@ public class FloorListModel : PageModel
 
     }
 
-    public void OnPostAddFloor()
+    public IActionResult OnPostAddFloor()
     {
-        SchoolId = int.Parse(HttpContext.Request.Cookies["SchoolId"]);
-        School = _schoolRepository.Get(SchoolId);
+        var sId = HttpContext.Request.Cookies["SchoolId"];
+        if (!int.TryParse(sId, out int schoolId))
+        {
+            return NotFound("School not found");
+        }
 
+        Floors = _floorRepository!.GetAll(f => f.SchoolId == schoolId);
+
+        var school = _schoolRepository.Get(schoolId);
         int number;
 
-        if (!Floors.Any() || Floors.Last().Number < 0)
+        if (!Floors!.Any() || Floors!.Last().Number < 0)
             number = 0;
         else
-            number = Floors.Last().Number;
+        {
+            number = Floors!.Last().Number;
+        }
 
-        Floor floor = new(number+1);
-        //_floorRepository!.Add(floor); ?
-        var (valid, error) = School.AddFloor(floor);
-        _ctx.SaveChanges();
+        Floor floor = new(number + 1);
+
+        floor.School = school!;
+
+        _floorRepository!.Add(floor);
+        return RedirectToPage("List");
     }
-    public void OnPostAddBasement()
+    public IActionResult OnPostAddBasement()
     {
-        SchoolId = int.Parse(HttpContext.Request.Cookies["SchoolId"]);
-        School = _schoolRepository.Get(SchoolId);
+        var sId = HttpContext.Request.Cookies["SchoolId"];
+        if (!int.TryParse(sId, out int schoolId))
+        {
+            return NotFound("School not found");
+        }
 
+        Floors = _floorRepository!.GetAll(f => f.SchoolId == schoolId);
+
+        var school = _schoolRepository.Get(schoolId);
         int number;
 
-        if (!Floors.Any() || Floors.First().Number >= 0)
+        if (!Floors!.Any() || Floors!.First().Number >= 0)
             number = 0;
         else
-            number = Floors.First().Number;
+            number = Floors!.First().Number;
 
         Floor floor = new(number - 1);
+        floor.School = school!;
 
-        var (valid, error) = School!.AddFloor(floor);
-        _ctx.SaveChanges();
+        _floorRepository!.Add(floor);
+        return RedirectToPage("List");
     }
     public IActionResult OnPostDelete(int id)
     {
-        var floor = _floorRepository.Get(id);
-        //_ctx.Remove(floor);
-        //_ctx.SaveChanges();
+        var floor = _floorRepository!.Get(id);
+
+        if (floor is null)
+        {
+            return NotFound("Floor not found");
+        }
+
         _floorRepository.Delete(floor!);
         return RedirectToPage("List");
     }
